requirements.txt
duckdb
python-dotenv
Jinja2
boto3
pandas
pyarrow
